+++ src/engine/link/pathInjector.tsx
@@ -1,5 +1,6 @@
-import { PathContext } from '@src/engine/link/context'
-import { FileRoute } from '../route/fileRoute'
+import { FileRoute } from '../route'
+
+import { PathContext } from './context'
 import { useAbsolutePath } from './useAbsolutePath'
 
 export const PathInjector: React.FC<{ route: Pick<FileRoute<string>, 'key'>; children: React.ReactNode }> = ({
