+++ tsconfig.json
@@ -21,6 +21,7 @@
     "paths": {
       "@src/*": ["./src/*"],
       "@data/*": ["./src/data/*"],
+      "@engine/*": ["./src/engine/*"],
       "@pages/*": ["./src/pages/*"],
       "@components/*": ["./src/components/*"]
     }
